Bug 776641: Import users provider from glarizza, plus my changes; r=kmoir
This is from
  https://github.com/glarizza/puppet-1/commit/1680efb087adfab96440398994e624aea5844592
adjusted to match 2.7.12, and renamed to the 'darwinuser' type.  The
contents of lib/puppet/provider/nameservice/directoryservice.rb were
included directly in lib/puppet/provider/darwinuser/directoryservice.rb,
with the class name adjusted to CustomDirectoryService.

My changes are most succinctly described at
  https://github.com/glarizza/puppet-1/pull/1

This code is *temporary* until some version of this makes it into a
released version of puppet, and PuppetAgain is upgraded to use that
version.