Bug 776641: Manage root and builder on OS X 10.8, take 2; r=kmoir
This splits account creation into child classes, both for clarity and
to allow them to be require'd by class name rather than by User[..].

This updates the darwinuser provider to the latest from glarizza:
  https://github.com/glarizza/puppet-1/commit/84621993be1b16e26fdf885d1a7315bb47fc9e6c

It adds the secrets and whatnot required to supply passwords for 10.8.

It deletes the 'administrator' account, which is an artifact of how we
set up these systems, but it can't hurt to ensure => absent.

This does *not* add the builder user to the Administrators group.  By
default, only Administrators are allowed to SSH to a machine, so the
builder user will not be permitted.  The sshd module in bug 764948 will
take care of this.

Finally, this adds setup/mtnlion-user-info.rb, for getting the password,
salt, and iterations values for existing users.

since the previous patch:
* remove screen resolution validation - doesn't work as root
* apply darwinuser provider updates from me and glarizza
* don't delete Administrator, for now, so we aren't locked out of hosts
* hack around race condition in creating new users