Bug 1055793 - Use runner on OS X; r=dustin
Installs runner along with a launchd init config which runs runner as the cltbld user.
Runner had previously assumed it would be run as root, by init.d/upstart, and
had opted to use 'su -' prior to starting the buildslave. Because starting
runner as cltbld is possible, and desirable for the resulting environment,
the Linux approach fails as it produces a user environment with no display and an incomplete path.
To fix this the runner buildbot task's template was modified to allow custom
buildslave commands to be set via puppet variables.