Bug 1114807 - Detect and remove empty/invalid pidfiles within the runner buildbot task; r=coop
An empty twistd.pid file results in an attempt by the task to convert '' to an int, which raises a ValueError.
The appropriate behavior here would be, as in the case of an OsError, to delete the empty file and move on.

--HG--
extra : rebase_source : 741c811df52f4f7bee8d2c414611368d3b232a41