Bug 1232442 - Seed images with a stream clone of the unified Firefox repo; r?catlee
The https://hg.mozilla.org/mozilla-unified repo is a single Mercurial
repository with relevant heads from all important repos (mozilla-central,
mozilla-inbound, mozilla-aurora, mozilla-release, esrs, etc). The repository
is encoded as generaldelta, which means it is smaller than mozilla-central
(even though it contains 30,000+ more commits!)

Recent work in automation (namely bug 1270317) changed automation to
always use shared, pooled storage for Mercurial repos. This meant
that we only need a single store for Firefox repos.

When this change was made, we didn't change AMI seeding. This means that
a worker would clone the Firefox repo on first job that needed it. This
is obviously inefficient.

This commit changes the shared repo seeding so the pooled/shared repo
now populated in automation is seeded at AMI generation time. So on
first job run, most commits will be present and we'll only do an
incremental pull. This restores the behavior from before bug 1270317
landed. There are multiple benefits:

1) Shared repo population will complete quicker (because we're only
   populating 1 repo)
2) We'll use less disk space for local repos (because we will only
   populate 1)
3) Jobs will start faster since most commits from most Firefox repos
   will already be present in the pre-populated shared repo.

The previous version of this file had code to map the instance's current
availability zone to an S3 bucket. As of bug 1249197, hg.mozilla.org
advertises bundle URLs to the appropriate S3 endpoint based on the
requesting IP. This favors same-AZ serving and means there should be 0
cost for data downloads from S3. Since this mapping is now done server
side as part of clone bundles, we remove this feature.

The previous version of this file downloaded a tar file of the .hg
directory for various repos and uncompressed it. The new version just
does an `hg clone` preferring "streaming clones." "Streaming clones" are
effectively `tar | nc` and are extremely fast. IMO the tar file provides
little value so it has been removed from the equation.

A downside of not using a tar file is that seeding now talks to
hg.mozilla.org instead of only S3. This could potentially drive a lot of
load to hg.mozilla.org if multiple machines perform this seeding at the
same time. However, 99+% of clone load will be offloaded to S3 via the
clone bundles and hg.mozilla.org will only need to serve commits since
the bundle was created. This should not be more than a few hundred
commits and should not require much effort on behalf of the server. But
if this does overwhelm the server, we can restore tar files (or
provide a clone mode that only fetches the bundle).

This commit assumes that all machines have Mercurial 3.7 as `hg` in
PATH. If an older version of Mercurial is present, the clone will take
several minutes longer than it should or it will fail due to the client
not having bundle2 support (the firefox repo requires bundle2).

A downside of this commit is that jobs not having the new
shared/pooled storage code deployed will need to perform a full clone
on first job because the old paths (e.g. /builds/hg-shared/mozilla-central)
are no longer present. This only impacts legacy commits/jobs and the
number of jobs should diminish over time. Furthermore, once hgtool
is updated to use shared/pooled storage, this won't be an issue (that
is tracked in bug 1270951).

MozReview-Commit-ID: 584WIULj5nQ

--HG--
extra : rebase_source : 1186e48ad43e9f965186a811a742f4deba3a8489