bug 1331098 - fix beetmover bustage. r=bustage
I updated a number of python packages.  One of them is chardet==3.0.2,
which doesn't contain this PR [1], which fixes a crash when we use
chardet with debug logging.  We don't directly call chardet, but
aiohttp does, so we're crashing during beetmover uploads.

Downgrading chardet until there's a new chardet>3.0.2 that fixes this issue.

[1] https://github.com/chardet/chardet/pull/117