use fqdn for ssl cert definition
The self-signed ssl certs are fqdn-specific, so it doesn't make a
lot of sense to allow for them to cross domains based on shared
hostname.

Also, `hiera_hash` appears to reference the global hiera secrets and
doesn't respect environments. Switching to `secret` instead appears to
work and also respect the environment.

This change assumes we're going to switch over the secrets at the same
time we land this patch set. If we want a more gradual change over, we
may want to introduce a new set of secrets (e.g.,
`signing_server_ssl_cert_new` ?) that can coexist with the old secrets.