reenable verbose for signing scriptworkers
Previously, we turned off verbosity for both signing scriptworkers
(daemon) and signingscript (script) to avoid so much log spam from the
signing scriptworkers.

However, the former a) broke papertrail logging at all for the signing
scriptworkers, since `verbose` specifies whether scriptworker logs to
the screen, and b) broke nagios monitoring, since the claimWork call is
what keeps the logfile up to date when the daemon is idle.

This patch reverts only the scriptworker daemon part of the patch, back
to being verbose.